backend/requirements.txt
flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
xrpl-py==2.4.0
click==8.1.7
python-dotenv==1.0.0
rich==13.7.0
requests==2.31.0 